Ported all remaining Condition tests to Integration tests.
