+++ tests/src/Integration/Condition/DataComparisonTest.php
@@ -30,6 +30,7 @@ class DataComparisonTest extends RulesIntegrationTestCase {
    */
   public function setUp() {
     parent::setUp();
+
     $this->condition = $this->conditionManager->createInstance('rules_data_comparison');
   }
