+++ tests/src/Integration/Condition/EntityHasFieldTest.php
@@ -27,6 +27,7 @@ class EntityHasFieldTest extends RulesEntityIntegrationTestCase {
    */
   public function setUp() {
     parent::setUp();
+
     $this->condition = $this->conditionManager->createInstance('rules_entity_has_field');
   }
