+++ tests/src/Integration/Condition/NodeIsStickyTest.php
@@ -2,20 +2,20 @@
 
 /**
  * @file
- * Contains \Drupal\Tests\rules\Unit\Condition\NodeIsPromotedTest.
+ * Contains \Drupal\Tests\rules\Integration\Condition\NodeIsStickyTest.
  */
 
-namespace Drupal\Tests\rules\Unit\Condition;
+namespace Drupal\Tests\rules\Integration\Condition;
 
 use Drupal\Core\Plugin\Context\ContextDefinition;
-use Drupal\rules\Plugin\Condition\NodeIsPromoted;
-use Drupal\Tests\rules\Unit\RulesUnitTestCase;
+use Drupal\rules\Plugin\Condition\NodeIsSticky;
+use Drupal\Tests\rules\Integration\RulesIntegrationTestCase;
 
 /**
- * @coversDefaultClass \Drupal\rules\Plugin\Condition\NodeIsPromoted
+ * @coversDefaultClass \Drupal\rules\Plugin\Condition\NodeIsSticky
  * @group rules_conditions
  */
-class NodeIsPromotedTest extends RulesUnitTestCase {
+class NodeIsStickyTest extends RulesIntegrationTestCase {
 
   /**
    * The condition to be tested.
@@ -30,11 +30,7 @@ class NodeIsPromotedTest extends RulesUnitTestCase {
   public function setUp() {
     parent::setUp();
 
-    $this->condition = new NodeIsPromoted([], '', ['context' => [
-      'node' => new ContextDefinition('entity:node'),
-    ]]);
-
-    $this->condition->setStringTranslation($this->getMockStringTranslation());
+    $this->condition = $this->conditionManager->createInstance('rules_node_is_sticky');
   }
 
   /**
@@ -43,7 +39,7 @@ class NodeIsPromotedTest extends RulesUnitTestCase {
    * @covers ::summary()
    */
   public function testSummary() {
-    $this->assertEquals('Node is promoted', $this->condition->summary());
+    $this->assertEquals('Node is sticky', $this->condition->summary());
   }
 
   /**
@@ -54,11 +50,11 @@ class NodeIsPromotedTest extends RulesUnitTestCase {
   public function testConditionEvaluation() {
     $node = $this->getMock('Drupal\node\NodeInterface');
     $node->expects($this->at(0))
-      ->method('isPromoted')
+      ->method('isSticky')
       ->will($this->returnValue(TRUE));
 
     $node->expects($this->at(1))
-      ->method('isPromoted')
+      ->method('isSticky')
       ->will($this->returnValue(FALSE));
 
     // Set the node context value.
